Moved all include files outside of src folder
