+++ include/ConsistentHash.hpp
