+++ include/swim/GossipFailureDetector.hpp
@@ -13,7 +13,7 @@
 #include "SwimServer.hpp"
 #include "SwimCommon.hpp"
 
-#include "utils/utils.hpp"
+#include "../utils/utils.hpp"
 
 
 namespace swim {
