+++ include/swim/SwimServer.hpp
@@ -11,7 +11,7 @@
 #include <thread>
 
 
-#include "utils/utils.hpp"
+#include "../utils/utils.hpp"
 #include "SwimCommon.hpp"
 #include "SwimClient.hpp"
