+++ src/swim/GossipFailureDetector.cpp
@@ -1,7 +1,7 @@
 // Copyright (c) 2017 AlertAvert.com. All rights reserved.
 // Created by M. Massenzio (marco@alertavert.com) on 2/12/17.
 
-#include "GossipFailureDetector.hpp"
+#include "../../include/swim/GossipFailureDetector.hpp"
 
 namespace swim {
