+++ src/utils/network.cpp
@@ -2,7 +2,7 @@
 // Created by M. Massenzio (marco@alertavert.com) on 10/9/16.
 
 
-#include "utils.hpp"
+#include "../../include/utils/utils.hpp"
 
 namespace utils {
